Split out ExternalResource
