+++ pontoon/base/models/external_resource.py
@@ -1,0 +1,22 @@
+from django.db import models
+
+from pontoon.base.models.locale import Locale
+from pontoon.base.models.project import Project
+
+
+class ExternalResource(models.Model):
+    """
+    Represents links to external project resources like staging websites,
+    production websites, development builds, production builds, screenshots,
+    langpacks, etc. or team resources like style guides, dictionaries,
+    glossaries, etc.
+    Has no relation to the Resource class.
+    """
+
+    locale = models.ForeignKey(Locale, models.CASCADE, blank=True, null=True)
+    project = models.ForeignKey(Project, models.CASCADE, blank=True, null=True)
+    name = models.CharField(max_length=32)
+    url = models.URLField("URL", blank=True)
+
+    def __str__(self):
+        return self.name
